Iniciando reestruturação do desacoplamento do projeto
