+++ lib/app_widget.dart
@@ -1,0 +1,23 @@
+import 'package:caderneta_campo_digital/pages/splash_screen.dart';
+import 'package:flutter/material.dart';
+import 'package:flutter_localizations/flutter_localizations.dart';
+
+class AppWidget extends StatelessWidget {
+  const AppWidget({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return MaterialApp(
+      title: 'Caderneta de Campo Digital',
+      debugShowCheckedModeBanner: false,
+      localizationsDelegates: GlobalMaterialLocalizations.delegates,
+      supportedLocales: const [
+        Locale('pt', 'BR'),
+      ],
+      theme: ThemeData(
+        primarySwatch: Colors.blue,
+      ),
+      home: SplashScreenPage(),
+    );
+  }
+}
