+++ lib/main.dart
@@ -1,27 +1,6 @@
+import 'package:caderneta_campo_digital/app_widget.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_localizations/flutter_localizations.dart';
-import 'package:caderneta_campo_digital/pages/splash_screen.dart';
 
 void main() {
-  runApp(MyApp());
-}
-
-class MyApp extends StatelessWidget {
-  const MyApp({Key? key}) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return MaterialApp(
-      title: 'Caderneta de Campo Digital',
-      debugShowCheckedModeBanner: false,
-      localizationsDelegates: GlobalMaterialLocalizations.delegates,
-      supportedLocales: const [
-        Locale('pt', 'BR'),
-      ],
-      theme: ThemeData(
-        primarySwatch: Colors.blue,
-      ),
-      home: SplashScreenPage(),
-    );
-  }
+  runApp(AppWidget());
 }
